Physical structure submodel removal et al
Includes also several other changes:
-Woody/ herbaceous now in survival submodel, feed into generation length
-Plant size now named ‘adult plant size’, now in survival submodel being
parent to adult survival and survival and abiotic disturbance response
-Genotypic plasticity/ diversity now partly assessed by number of sub
species and varieties
-Generation time/ age of first reproduction is a complicated trait:
short is favorable for migration and evolution but not for toleration:
revised model to simplify: removed intermediary age of first
reproduction as that is entirely related/ dependent on generation
length. This factor now influences survival (which impacts toleration)
and reproduction (which impacts migrate and evolve response).
-Evolutionary history: removed favorable/ unfavorable. Define states for
several other factors less subjectively